Rename namespace to DumpAnalysisService
