+++ DumpAnalysisService.Tests/PathDetectionServiceTests.cs
@@ -1,9 +1,9 @@
-using BackgroundService.Infrastructure.Detection;
-using BackgroundService.Infrastructure.IO;
+using DumpAnalysisService.Infrastructure.Detection;
+using DumpAnalysisService.Infrastructure.IO;
 using Microsoft.Extensions.Logging;
 using Moq;
 
-namespace BackgroundService.Tests;
+namespace DumpAnalysisService.Tests;
 
 public class PathDetectionServiceTests
 {
