+++ DumpAnalysisService/Controllers/DiagnosticsController.cs
@@ -1,10 +1,10 @@
 using System.Reflection;
-using BackgroundService.Services;
+using DumpAnalysisService.Services;
 using Microsoft.AspNetCore.Mvc;
 using Shared;
 using Shared.Models;
 
-namespace BackgroundService.Controllers;
+namespace DumpAnalysisService.Controllers;
 
 [ApiController]
 [Route("api/diagnostics")]
