+++ DumpAnalysisService/Models/BasicAnalysisRequest.cs
@@ -1,3 +1,3 @@
-namespace BackgroundService.Models;
+namespace DumpAnalysisService.Models;
 
 public record BasicAnalysisRequest(string SessionId);
