+++ DumpAnalysisService/Models/ExecuteCommandRequest.cs
@@ -1,3 +1,3 @@
-namespace BackgroundService.Models;
+namespace DumpAnalysisService.Models;
 
 public record ExecuteCommandRequest(string SessionId, string Command);
