+++ DumpAnalysisService/Models/LoadDumpRequest.cs
@@ -1,3 +1,3 @@
-namespace BackgroundService.Models;
+namespace DumpAnalysisService.Models;
 
 public record LoadDumpRequest(string DumpFilePath);
