+++ DumpAnalysisService/Models/PredefinedAnalysisRequest.cs
@@ -1,3 +1,3 @@
-namespace BackgroundService.Models;
+namespace DumpAnalysisService.Models;
 
 public record PredefinedAnalysisRequest(string SessionId, string AnalysisType);
